refactor: split hooks into two files
