+++ packages/shared/src/features/opportunity/hooks/useOpportunity.ts
@@ -2,22 +2,10 @@ import { useQuery } from '@tanstack/react-query';
 import type { QueryKey } from '@tanstack/react-query';
 import { RequestKey, StaleTime } from '../../../lib/query';
 import { gqlClient } from '../../../graphql/common';
-import {
-  GET_OPPORTUNITY_MATCH_QUERY,
-  OPPORTUNITY_BY_ID_QUERY,
-} from '../graphql';
+import { OPPORTUNITY_BY_ID_QUERY } from '../graphql';
 import type { Organization } from '../../organizations/types';
 import type { PublicProfile } from '../../../lib/user';
 import type { ProtoEnumValue } from '../../../lib/protobuf';
-
-export enum OpportunityMatchStatus {
-  Pending = 'pending',
-  CandidateAccepted = 'candidate_accepted',
-  CandidateRejected = 'candidate_rejected',
-  CandidateTimeOut = 'candidate_time_out',
-  RecruiterAccepted = 'recruiter_accepted',
-  RecruiterRejected = 'recruiter_rejected',
-}
 
 export const getOpportunityByIdKey = (id: string): QueryKey => [
   RequestKey.Opportunity,
@@ -67,13 +55,6 @@ export type Opportunity = {
   }[];
 };
 
-export type OpportunityMatch = {
-  status: OpportunityMatchStatus;
-  description?: {
-    reasoning: string;
-  };
-};
-
 export const opportunityByIdOptions = ({ id }: { id: string }) => {
   return {
     queryKey: getOpportunityByIdKey(id),
@@ -91,33 +72,10 @@ export const opportunityByIdOptions = ({ id }: { id: string }) => {
   };
 };
 
-export const opportunityMatchOptions = ({ id }: { id: string }) => {
-  return {
-    queryKey: [...getOpportunityByIdKey(id), 'match'],
-    queryFn: async () => {
-      const res = await gqlClient.request<{
-        getOpportunityMatch: OpportunityMatch;
-      }>(GET_OPPORTUNITY_MATCH_QUERY, {
-        id,
-      });
-
-      return res.getOpportunityMatch;
-    },
-    staleTime: StaleTime.Default,
-    enabled: !!id,
-  };
-};
-
 export const useOpportunity = (id: string) => {
   const { data: opportunity, isPending } = useQuery(
     opportunityByIdOptions({ id }),
   );
 
   return { opportunity, isPending };
-};
-
-export const useOpportunityMatch = (id: string) => {
-  const { data: match } = useQuery(opportunityMatchOptions({ id }));
-
-  return { match };
 };
